(PC-8337) Show ended bookings even if no booking
