+++ src/features/bookings/components/OnGoingBookingsList.tsx
@@ -1,16 +1,13 @@
 import { plural } from '@lingui/macro'
-import { useNavigation } from '@react-navigation/native'
 import React, { useCallback } from 'react'
-import { FlatList, ListRenderItem, NativeScrollEvent, View } from 'react-native'
+import { FlatList, ListRenderItem, NativeScrollEvent } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import styled from 'styled-components/native'
 
-import { UseNavigationType } from 'features/navigation/RootNavigator'
+import { EndedBookingsSection } from 'features/bookings/pages/EndedBookingsSection'
 import { useFunctionOnce } from 'features/offer/services/useFunctionOnce'
 import { analytics } from 'libs/analytics'
 import { isCloseToBottom } from 'libs/analytics.utils'
-import { Badge } from 'ui/components/Badge'
-import { SectionRow } from 'ui/components/SectionRow'
 import { TAB_BAR_COMP_HEIGHT } from 'ui/theme'
 import { ColorsEnum, getSpacing, Typo } from 'ui/theme'
 
@@ -26,8 +23,6 @@ interface OnGoingBookingsListProps {
 const emptyBookings: Booking[] = []
 
 export function OnGoingBookingsList(props: OnGoingBookingsListProps) {
-  const { navigate } = useNavigation<UseNavigationType>()
-
   const { bottom } = useSafeAreaInsets()
 
   const bookings = props.bookings || emptyBookings
@@ -38,35 +33,18 @@ export function OnGoingBookingsList(props: OnGoingBookingsListProps) {
     other: '# réservations en cours',
   })
 
-  const endedBookings = props?.endedBookings || emptyBookings
-  const endedBookingsLabel = plural(endedBookings.length, {
-    one: 'Réservation terminée',
-    other: 'Réservations terminées',
-  })
-
   const ListEmptyComponent = useCallback(() => <NoBookingsView />, [])
   const ListHeaderComponent = useCallback(
     () => (hasBookings ? <BookingsCount>{bookingsCountLabel}</BookingsCount> : null),
     [hasBookings, bookingsCountLabel]
   )
   const ListFooterComponent = useCallback(
-    () =>
-      endedBookings.length > 0 ? (
-        <EndedBookingsSection safeBottom={bottom}>
-          <View>
-            <SectionRow
-              type="navigable"
-              title={endedBookingsLabel}
-              icon={() => <Badge label={endedBookings.length} />}
-              onPress={() => navigate('EndedBookings')}
-              testID="row-ended-bookings"
-            />
-          </View>
-        </EndedBookingsSection>
-      ) : (
-        <React.Fragment />
-      ),
-    [endedBookings, bottom, endedBookingsLabel]
+    () => (
+      <FooterContainer safeBottom={bottom}>
+        <EndedBookingsSection />
+      </FooterContainer>
+    ),
+    [hasBookings]
   )
 
   const logBookingsScrolledToBottom = useFunctionOnce(analytics.logBookingsScrolledToBottom)
@@ -119,9 +97,7 @@ const BookingsCount = styled(Typo.Body).attrs({
   paddingBottom: getSpacing(2),
 })
 
-const EndedBookingsSection = styled.View<{ safeBottom: number }>(({ safeBottom }) => ({
-  width: '100%',
+const FooterContainer = styled.View<{ safeBottom: number }>(({ safeBottom }) => ({
   marginBottom: safeBottom ? safeBottom / 2 : 0,
-  justifyContent: 'center',
   paddingVertical: getSpacing(4),
 }))
